move getBpmnIsInstantiating in style/utils.ts where it was previously stored.
This fix test isses.
