+++ src/component/mxgraph/style/index.ts
@@ -14,5 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-export * from './utils';
+export { StyleDefault } from './utils';
 export * from './identifiers';
